Submissions admin fixes (#1023)
Closes
https://github.com/SoftUni-Internal/exam-systems-issues/issues/1150

---------

+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/administration/common/delete/DeleteButton.tsx
@@ -18,7 +18,13 @@ interface IDeleteButtonProps {
 }
 
 const DeleteButton = (props: IDeleteButtonProps) => {
-    const { onSuccess, id, name, style = {}, text, mutation } = props;
+    const {
+        onSuccess,
+        id,
+        name,
+        style = {},
+        text, mutation,
+    } = props;
     const [ showConfirmDelete, setShowConfirmDelete ] = useState<boolean>(false);
     const [ message, setMessage ] = useState<string | null>(null);
     const [ errorMessages, setErrorMessages ] = useState<Array<string>>([]);
@@ -59,15 +65,15 @@ const DeleteButton = (props: IDeleteButtonProps) => {
                             <IoMdTrash color="red" />
                         </IconButton>
                     </Tooltip>
-                    {showConfirmDelete && (
-                    <ConfirmDialog
-                      title={`Delete: ${name}`}
-                      text={text}
-                      confirmButtonText="Delete"
-                      declineButtonText="Cancel"
-                      onClose={() => setShowConfirmDelete(!showConfirmDelete)}
-                      confirmFunction={() => deleteRequest(id)}
-                    />
+                    { showConfirmDelete && (
+                        <ConfirmDialog
+                          title={`Delete: ${name}`}
+                          text={text}
+                          confirmButtonText="Delete"
+                          declineButtonText="Cancel"
+                          onClose={() => setShowConfirmDelete(!showConfirmDelete)}
+                          confirmFunction={() => deleteRequest(id)}
+                        />
                     )}
                 </div>
             )
